third_party/WebKit/LayoutTests/editing/input/text-input-controller-leak-document.html

<html>
<body>
    <input id='text' type='text'></input>

    <script src="../../resources/js-test.js"></script>
    <script>
        var input = document.getElementById('text');
        input.focus();
        textInputController.setMarkedText("hello", 0, 5);
        // Do not pass the object directly to observeGC function. This may
        // remain live on this function's stack preventing GC from collecting
        // it. Accessing the object inside an inner function will prevent any
        // unneeded references on this function's stack.
        var markedRangeResultGC = internals.observeGC((() => {
                return textInputController.markedRange();
            })());
        var selectedRangeResultGC = internals.observeGC((() => {
                return textInputController.selectedRange();
            })());
        var firstRectForCharacterRangeGC = internals.observeGC((() => {
                return textInputController.firstRectForCharacterRange(0, 0);
            })());
        gc();
        shouldBeTrue('markedRangeResultGC.wasCollected');
        shouldBeTrue('selectedRangeResultGC.wasCollected');
        shouldBeTrue('firstRectForCharacterRangeGC.wasCollected');
    </script>
</body>
</html>
